Flytter bønne for å unngå sirkulær bønneopprettelse
